{"schema":"decodex/commit/1","summary":"Split MCP prompt modules","authority":"XY-1115"}
