+++ apps/decodex/src/mcp/prompts/arguments.rs
@@ -1,0 +1,15 @@
+use serde_json::Value;
+
+pub(super) fn prompt_argument<'a>(arguments: &'a Value, key: &str) -> Option<&'a str> {
+	arguments.get(key).and_then(Value::as_str).map(str::trim).filter(|value| !value.is_empty())
+}
+
+pub(super) fn prompt_required_arguments_are_present(name: &str, arguments: &Value) -> bool {
+	let required: &[&str] = match name {
+		"decodex_research" => &["intent"],
+		"decodex_validation_ready" | "decodex_handoff" | "decodex_lane_control" => &["issue"],
+		_ => return true,
+	};
+
+	required.iter().all(|key| prompt_argument(arguments, key).is_some())
+}
