+++ apps/decodex/src/mcp/prompts/catalog.rs
@@ -1,0 +1,64 @@
+use serde_json::Value;
+
+pub(super) fn mcp_prompts() -> Vec<Value> {
+	vec![
+		serde_json::json!({
+			"name": "decodex_research",
+			"title": "Decodex Research",
+			"description": "Frame bounded Decodex research as a latent Decision Contract candidate.",
+			"arguments": [
+				{
+					"name": "intent",
+					"description": "Natural-language research question or design uncertainty.",
+					"required": true
+				}
+			]
+		}),
+		serde_json::json!({
+			"name": "decodex_validation_ready",
+			"title": "Decodex Validation Ready",
+			"description": "Drive an implementation or repair lane to local validation-ready evidence.",
+			"arguments": [
+				{
+					"name": "issue",
+					"description": "Linear issue identifier for the lane.",
+					"required": true
+				},
+				{
+					"name": "phase",
+					"description": "Current Decodex phase goal.",
+					"required": false
+				}
+			]
+		}),
+		serde_json::json!({
+			"name": "decodex_handoff",
+			"title": "Decodex Handoff",
+			"description": "Prepare a verified review handoff only after local validation and bounded review.",
+			"arguments": [
+				{
+					"name": "issue",
+					"description": "Linear issue identifier for the lane.",
+					"required": true
+				}
+			]
+		}),
+		serde_json::json!({
+			"name": "decodex_lane_control",
+			"title": "Decodex Lane Control",
+			"description": "Inspect first, then request guarded lane-control actions through existing Decodex authority gates.",
+			"arguments": [
+				{
+					"name": "issue",
+					"description": "Linear issue identifier or local tracker issue id.",
+					"required": true
+				},
+				{
+					"name": "runId",
+					"description": "Current run id observed through lane inspect.",
+					"required": false
+				}
+			]
+		}),
+	]
+}
